COMMERCE-1950 talend - component - move json related code to specialized classes and free SourceOrSink implementation of unrelated code
