+++ modules/etl/talend/talend-common/src/main/java/com/liferay/talend/common/oas/OASSource.java
@@ -1,0 +1,31 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.talend.common.oas;
+
+import javax.json.JsonObject;
+
+import org.talend.components.api.properties.ComponentProperties;
+import org.talend.daikon.properties.ValidationResult;
+
+/**
+ * @author Igor Beslic
+ */
+public interface OASSource {
+
+	public JsonObject getOASJsonObject(String oasURL);
+
+	public ValidationResult initialize(ComponentProperties componentProperties);
+
+}
