+++ modules/etl/talend/talend-runtime/src/main/java/com/liferay/talend/runtime/LiferaySourceOrSink.java
@@ -14,12 +14,8 @@
 
 package com.liferay.talend.runtime;
 
-import com.liferay.talend.avro.EndpointSchemaInferrer;
 import com.liferay.talend.common.exception.MalformedURLException;
-import com.liferay.talend.common.json.JsonFinder;
-import com.liferay.talend.common.oas.OASParameter;
-import com.liferay.talend.common.oas.constants.OASConstants;
-import com.liferay.talend.common.util.StringUtil;
+import com.liferay.talend.common.oas.OASSource;
 import com.liferay.talend.common.util.URIUtil;
 import com.liferay.talend.connection.LiferayConnectionProperties;
 import com.liferay.talend.connection.LiferayConnectionPropertiesProvider;
@@ -30,15 +26,8 @@ import com.liferay.talend.runtime.client.exception.ResponseContentClientExceptio
 
 import java.io.IOException;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
-import java.util.TreeMap;
-import java.util.TreeSet;
 
-import javax.json.JsonArray;
 import javax.json.JsonObject;
 
 import javax.ws.rs.ProcessingException;
@@ -67,8 +56,7 @@ import org.talend.daikon.properties.ValidationResultMutable;
  * @author Ivica Cardic
  */
 public class LiferaySourceOrSink
-	extends TranslatableImpl
-	implements LiferaySourceOrSinkRuntime, SourceOrSink {
+	extends TranslatableImpl implements OASSource, SourceOrSink {
 
 	public JsonObject doDeleteRequest(RuntimeContainer runtimeContainer) {
 		return doDeleteRequest(runtimeContainer, null);
@@ -202,39 +190,6 @@ public class LiferaySourceOrSink
 		return liferayConnectionProperties;
 	}
 
-	@Override
-	public Set<String> getEndpointList(String operation) {
-		LiferayConnectionProperties liferayConnectionProperties =
-			getEffectiveConnection(null);
-
-		JsonObject oasJsonObject = doGetRequest(
-			liferayConnectionProperties.getApiSpecURL());
-
-		return _extractEndpoints(operation, oasJsonObject);
-	}
-
-	@Override
-	public Map<String, String> getEndpointMap(String operation) {
-		Map<String, String> endpointMap = new TreeMap<>();
-
-		LiferayConnectionProperties liferayConnectionProperties =
-			getEffectiveConnection(null);
-
-		JsonObject oasJsonObject = doGetRequest(
-			liferayConnectionProperties.getApiSpecURL());
-
-		Set<String> endpoints = _extractEndpoints(operation, oasJsonObject);
-
-		for (String endpoint : endpoints) {
-			endpointMap.put(
-				endpoint,
-				_endpointSchemaInferrer.extractEndpointSchemaName(
-					endpoint, operation, oasJsonObject));
-		}
-
-		return endpointMap;
-	}
-
 	/**
 	 * @deprecated As of Mueller (7.2.x)
 	 */
@@ -247,41 +202,8 @@ public class LiferaySourceOrSink
 	}
 
 	@Override
-	public Schema getEndpointSchema(String endpoint, String operation) {
-		LiferayConnectionProperties liferayConnectionProperties =
-			getEffectiveConnection(null);
-
-		JsonObject oasJsonObject = doGetRequest(
-			liferayConnectionProperties.getApiSpecURL());
-
-		return _endpointSchemaInferrer.inferSchema(
-			endpoint, operation, oasJsonObject);
-	}
-
-	@Override
-	public List<OASParameter> getParameters(String endpoint, String operation) {
-		List<OASParameter> oasParameters = new ArrayList<>();
-
-		LiferayConnectionProperties liferayConnectionProperties =
-			getEffectiveConnection(null);
-
-		String apiSpecURLHref = liferayConnectionProperties.getApiSpecURL();
-
-		JsonObject oasJsonObject = doGetRequest(apiSpecURLHref);
-
-		String jsonFinderPath = StringUtil.replace(
-			OASConstants.PATH_ENDPOINT_OPERATION_PARAMETERS_PATTERN,
-			"ENDPOINT_TPL", endpoint, "OPERATION_TPL", operation);
-
-		JsonArray parametersJsonArray = _jsonFinder.getDescendantJsonArray(
-			jsonFinderPath, oasJsonObject);
-
-		for (int i = 0; i < parametersJsonArray.size(); i++) {
-			oasParameters.add(
-				_toParameter(parametersJsonArray.getJsonObject(i)));
-		}
-
-		return oasParameters;
+	public JsonObject getOASJsonObject(String oasURL) {
+		return doGetRequest(null, oasURL);
 	}
 
 	public RESTClient getRestClient(RuntimeContainer runtimeContainer) {
@@ -321,20 +243,18 @@ public class LiferaySourceOrSink
 		throw new UnsupportedOperationException();
 	}
 
-	public Set<String> getSupportedOperations(String endpoint) {
-		LiferayConnectionProperties liferayConnectionProperties =
-			getEffectiveConnection(null);
+	@Override
+	public ValidationResult initialize(
+		ComponentProperties componentProperties) {
 
-		JsonObject oasJsonObject = doGetRequest(
-			liferayConnectionProperties.getApiSpecURL());
+		ValidationResult validationResult = initialize(
+			null, componentProperties);
 
-		String jsonFinderPath = StringUtil.replace(
-			OASConstants.PATH_ENDPOINT_PATTERN, "ENDPOINT_TPL", endpoint);
+		if (validationResult.getStatus() == ValidationResult.Result.ERROR) {
+			return validationResult;
+		}
 
-		JsonObject endpointJsonObject = _jsonFinder.getDescendantJsonObject(
-			jsonFinderPath, oasJsonObject);
-
-		return endpointJsonObject.keySet();
+		return validate(null);
 	}
 
 	@Override
@@ -401,13 +321,10 @@ public class LiferaySourceOrSink
 					"error.validation.connection.credentials"));
 		}
 
-		return validateConnection(
-			liferayConnectionProperties, runtimeContainer);
+		return validateConnection(runtimeContainer);
 	}
 
-	@Override
 	public ValidationResult validateConnection(
-		LiferayConnectionPropertiesProvider liferayConnectionPropertiesProvider,
 		RuntimeContainer runtimeContainer) {
 
 		try {
@@ -462,49 +379,6 @@ public class LiferaySourceOrSink
 		liferayConnectionPropertiesProvider;
 	protected RESTClient restClient;
 
-	private Set<String> _extractEndpoints(
-		String operation, JsonObject oasJsonObject) {
-
-		Set<String> endpoints = new TreeSet<>();
-
-		JsonObject pathsJsonObject = oasJsonObject.getJsonObject(
-			OASConstants.PATHS);
-
-		pathsJsonObject.forEach(
-			(path, operationsJsonValue) -> {
-				JsonObject operationsJsonObject =
-					operationsJsonValue.asJsonObject();
-
-				operationsJsonObject.forEach(
-					(operationName, operationJsonValue) -> {
-						if (!Objects.equals(operation, operationName)) {
-							return;
-						}
-
-						if (!Objects.equals(
-								operation, OASConstants.OPERATION_GET)) {
-
-							endpoints.add(path);
-
-							return;
-						}
-
-						if (_jsonFinder.hasPath(
-								OASConstants.PATH_RESPONSE_SCHEMA_REFERENCE,
-								operationJsonValue.asJsonObject()) ||
-							_jsonFinder.hasPath(
-								OASConstants.
-									PATH_RESPONSE_SCHEMA_ITEMS_REFERENCE,
-								operationJsonValue.asJsonObject())) {
-
-							endpoints.add(path);
-						}
-					});
-			});
-
-		return endpoints;
-	}
-
 	private boolean _isNullString(String value) {
 		if (value == null) {
 			return true;
@@ -519,25 +393,11 @@ public class LiferaySourceOrSink
 		return false;
 	}
 
-	private OASParameter _toParameter(JsonObject jsonObject) {
-		OASParameter oasParameter = new OASParameter(
-			jsonObject.getString("name"), jsonObject.getString("in"));
-
-		if (jsonObject.containsKey("required")) {
-			oasParameter.setRequired(jsonObject.getBoolean("required"));
-		}
-
-		return oasParameter;
-	}
-
 	private static final Logger _logger = LoggerFactory.getLogger(
 		LiferaySourceOrSink.class);
 
 	private static final long serialVersionUID = 3109815759807236523L;
 
-	private final EndpointSchemaInferrer _endpointSchemaInferrer =
-		new EndpointSchemaInferrer();
-	private final JsonFinder _jsonFinder = new JsonFinder();
 	private final ResponseHandler _responseHandler = new ResponseHandler();
 
 }
